Use ComposeCanvases component for compose logic
