+++ visualizer/src/Canvas/Canvas.js
@@ -2,8 +2,9 @@ import Box from '@material-ui/core/Box';
 import CircularProgress from '@material-ui/core/CircularProgress';
 import { makeStyles } from '@material-ui/core/styles';
 import { useSelector } from '@xstate/react';
-import React, { useEffect, useRef } from 'react';
+import React, { useEffect } from 'react';
 import { useCanvas, useLabeled, useRaw, useSegment, useSelect } from '../ProjectContext';
+import ComposeCanvas from './ComposeCanvases';
 import LabeledCanvas from './Labeled/LabeledCanvas';
 import OutlineCanvas from './Labeled/OutlineCanvas';
 import RawCanvas from './Raw/RawCanvas';
@@ -12,13 +13,6 @@ const useStyles = makeStyles({
   canvasBox: {
     alignSelf: 'flex-start',
     position: 'absolute',
-  },
-  canvas: {
-    position: 'absolute',
-    top: 0,
-    left: 0,
-    maxHeight: '100%',
-    maxWidth: '100%',
   },
 });
 
@@ -35,15 +29,11 @@ export const Canvas = () => {
   const sh = useSelector(canvas, (state) => state.context.height);
   const scale = useSelector(canvas, (state) => state.context.scale);
 
-  const width = sw * scale * window.devicePixelRatio;
-  const height = sh * scale * window.devicePixelRatio;
-
   const grab = useSelector(canvas, (state) => state.matches('pan.grab'));
   const grabbing = useSelector(canvas, (state) => state.matches('pan.grab.panning'));
   const dragged = useSelector(canvas, (state) =>
     state.matches('pan.interactive.panOnDrag.dragged')
   );
-
   const cursor = grabbing || dragged ? 'grabbing' : grab ? 'grab' : 'crosshair';
 
   const segment = useSegment();
@@ -91,70 +81,7 @@ export const Canvas = () => {
     }
   };
 
-  const canvasRef = useRef();
-  const ctxRef = useRef();
-
-  const composeCanvasRef = useRef();
-  const composeCtxRef = useRef();
-
-  useEffect(() => {
-    composeCtxRef.current = composeCanvasRef.current.getContext('2d');
-    composeCtxRef.current.globalCompositeOperation = 'screen';
-  }, [sh, sw]);
-
-  useEffect(() => {
-    ctxRef.current = canvasRef.current.getContext('2d');
-    ctxRef.current.imageSmoothingEnabled = false;
-  }, [height, width]);
-
-  // const render = () => {
-  //   composeCtxRef.current.clearRect(0, 0, sw, sh);
-  //   for (let key in canvases) {
-  //     if (key !== 'key') {
-  //       composeCtxRef.current.drawImage(
-  //         canvases[key],
-  //         sx,
-  //         sy,
-  //         sw / zoom,
-  //         sh / zoom,
-  //         0,
-  //         0,
-  //         width,
-  //         height
-  //       );
-  //     }
-  //   }
-  //   console.log('Composed canvas: ', canvases);
-  //   // forceUpdate();
-  // };
-
   const [canvases, setCanvases] = React.useState({});
-
-  useEffect(() => {
-    composeCtxRef.current.clearRect(0, 0, sw, sh);
-    for (let key in canvases) {
-      if (key !== 'key') {
-        composeCtxRef.current.drawImage(canvases[key], 0, 0);
-      }
-    }
-    console.log('Composed canvases: ', canvases);
-  }, [canvases, sh, sw]);
-
-  useEffect(() => {
-    ctxRef.current.clearRect(0, 0, width, height);
-    ctxRef.current.drawImage(
-      composeCanvasRef.current,
-      sx,
-      sy,
-      sw / zoom,
-      sh / zoom,
-      0,
-      0,
-      width,
-      height
-    );
-    console.log('Rendered canvas');
-  }, [canvases, sx, sy, sw, sh, zoom, width, height]);
 
   return (
     <Box
@@ -170,9 +97,7 @@ export const Canvas = () => {
       onMouseUp={canvas.send}
     >
       {!raw && <CircularProgress style={{ margin: '25%', width: '50%', height: '50%' }} />}
-      {/* <ComposeCanvas canvases={canvases} /> */}
-      <canvas id='compose-canvas' hidden={true} ref={composeCanvasRef} width={sw} height={sh} />
-      <canvas id='canvas' className={styles.canvas} ref={canvasRef} width={width} height={height} />
+      <ComposeCanvas canvases={canvases} />
       {raw && <RawCanvas setCanvases={setCanvases} />}
       {labeled && <LabeledCanvas setCanvases={setCanvases} />}
       {labeled && <OutlineCanvas setCanvases={setCanvases} />}
