+++ visualizer/src/Canvas/ComposeCanvases.js
@@ -1,9 +1,20 @@
+import { makeStyles } from '@material-ui/core';
 import { useSelector } from '@xstate/react';
 import React, { useEffect, useRef } from 'react';
-import { useCanvas, useSelect } from '../ProjectContext';
+import { useCanvas } from '../ProjectContext';
 
-export const ComposeCanvas = ({ children }) => {
-  const select = useSelect();
+const useStyles = makeStyles({
+  canvas: {
+    position: 'absolute',
+    top: 0,
+    left: 0,
+    maxHeight: '100%',
+    maxWidth: '100%',
+  },
+});
+
+export const ComposeCanvas = ({ canvases }) => {
+  const styles = useStyles();
 
   const canvas = useCanvas();
   const sx = useSelector(canvas, (state) => state.context.sx);
@@ -13,67 +24,52 @@ export const ComposeCanvas = ({ children }) => {
   const sh = useSelector(canvas, (state) => state.context.height);
   const scale = useSelector(canvas, (state) => state.context.scale);
 
-  // prevent scrolling page when over canvas
-  useEffect(() => {
-    const composeCanvas = document.getElementById('composeCanvas');
-    const wheelListener = (e) => e.preventDefault();
-    const spaceListener = (e) => {
-      if (e.key === ' ') {
-        e.preventDefault();
-      }
-    };
-    composeCanvas.addEventListener('wheel', wheelListener);
-    document.addEventListener('keydown', spaceListener);
-    return () => {
-      composeCanvas.removeEventListener('wheel', wheelListener);
-      document.removeEventListener('keydown', spaceListener);
-    };
-  }, []);
-
-  const handleMouseDown = (event) => {
-    event.preventDefault();
-    if (event.shiftKey) {
-      select.send({ ...event, type: 'SHIFT_CLICK' });
-    } else {
-      canvas.send(event);
-    }
-  };
-
   const width = sw * scale * window.devicePixelRatio;
   const height = sh * scale * window.devicePixelRatio;
 
   const canvasRef = useRef();
-  const ctx = useRef();
+  const ctxRef = useRef();
+
+  const composeCanvasRef = useRef();
+  const composeCtxRef = useRef();
 
   useEffect(() => {
-    ctx.current = canvasRef.current.getContext('2d');
-    ctx.current.imageSmoothingEnabled = false;
-    ctx.current.globalCompositeOperation = 'source-over'; // default
-  }, [width, height]);
+    composeCtxRef.current = composeCanvasRef.current.getContext('2d');
+    composeCtxRef.current.globalCompositeOperation = 'screen';
+  }, [sh, sw]);
 
   useEffect(() => {
-    ctx.current.save();
-    ctx.current.clearRect(0, 0, width, height);
-    for (const child of children) {
-      console.log(child, child instanceof HTMLCanvasElement);
-      if (child instanceof HTMLCanvasElement) {
-        ctx.current.drawImage(child, sx, sy, sw / zoom, sh / zoom, 0, 0, width, height);
-      }
+    ctxRef.current = canvasRef.current.getContext('2d');
+    ctxRef.current.imageSmoothingEnabled = false;
+  }, [height, width]);
+
+  useEffect(() => {
+    composeCtxRef.current.clearRect(0, 0, sw, sh);
+    for (let key in canvases) {
+      composeCtxRef.current.drawImage(canvases[key], 0, 0);
     }
-  }, [children, sx, sy, sw, sh, zoom, width, height]);
+  }, [canvases, sh, sw]);
+
+  useEffect(() => {
+    ctxRef.current.clearRect(0, 0, width, height);
+    ctxRef.current.drawImage(
+      composeCanvasRef.current,
+      sx,
+      sy,
+      sw / zoom,
+      sh / zoom,
+      0,
+      0,
+      width,
+      height
+    );
+  }, [canvases, sx, sy, sw, sh, zoom, width, height]);
 
   return (
-    <canvas
-      id={'composeCanvas'}
-      boxShadow={10}
-      width={width}
-      height={height}
-      onMouseMove={canvas.send}
-      onWheel={canvas.send}
-      onMouseDown={handleMouseDown}
-      onMouseUp={canvas.send}
-      ref={canvasRef}
-    />
+    <>
+      <canvas id='compose-canvas' hidden={true} ref={composeCanvasRef} width={sw} height={sh} />
+      <canvas id='canvas' className={styles.canvas} ref={canvasRef} width={width} height={height} />
+    </>
   );
 };
